Lock CocoaPods dependencies' versions
Until now, none of the dependencies had their version locked in either
the .podspec or Podfile. So, if one of them got updated with breaking
changes then building would ouviously break after a `pod update`.

This issue happened recently when Bolts-Swift was updated with breaking
changes from 1.1.0 to 1.2.0.

This addresses that by locking the dependency versions to the latest
compatible version at the moment. When support is added to the new
dependency versions, the podspec/Podfile should be updated and the pod
version should be bumped.